backend/requirements.txt


# Core Python packages
python-dotenv>=1.0.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-dateutil>=2.9.0.post0
pytz>=2024.1

# Data processing and scientific computing (Latest versions with Python 3.13 support)
numpy>=2.3.0
pandas>=2.3.0
scipy>=1.13.1
scikit-learn>=1.7.0
statsmodels>=0.14.2

# Database
sqlalchemy>=2.0.30
alembic>=1.13.2
databases>=0.9.0
psycopg2-binary>=2.9.9

# Web frameworks
Django>=5.2.3
django-cors-headers>=4.4.0
django-filter>=24.2
django-extensions>=3.2.3
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.2
drf-yasg>=1.21.7

fastapi>=0.111.0
uvicorn[standard]>=0.30.1
gunicorn>=22.0.0

# Data visualization
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.22.0

# Machine Learning
xgboost>=2.1.0
lightgbm>=4.4.0

# NLP/Text analysis
nltk>=3.8.1
textblob>=0.18.0
wordcloud>=1.9.3
spacy>=3.7.5

# Utilities
requests>=2.32.3
httpx>=0.27.0
aiofiles>=23.2.1
python-multipart>=0.0.9
pillow>=10.3.0

# Security
cryptography>=42.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Task Queue & Caching
celery>=5.4.0
redis>=5.0.6
django-redis>=5.4.0

# Storage
django-storages>=1.14.4
boto3>=1.34.131

# Logging
loguru>=0.7.2

# Testing
pytest>=8.2.2
pytest-django>=4.8.0
pytest-asyncio>=0.23.7
factory-boy>=3.3.0

# Development tools
black>=24.4.2
flake8>=7.1.0
mypy>=1.10.0
